requirements.txt
python-dotenv
openai
deepeval
